MySensors: add helper to find MySensors nodes using TCP
In order to detect MySensqors nodes, use a simple mysensors discovery.
This discovery is done by sending a presentation request to all nodes.
If the node is present, it will send its sensors. Some older MySensors
version did not send their sensors on connection so sensors might be
presented twice by new nodes. This is filtered using a temporary list
for the discovery and then display the detected sensors.